ml_training/scripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0
scikit-learn>=1.3.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
numpy>=1.24.0
